+++ bld/sdk/wre/c/wre_wres.c
@@ -2,6 +2,7 @@
 *
 *                            Open Watcom Project
 *
+* Copyright (c) 2002-2022 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -49,7 +50,7 @@
 WResID *WRECopyWResID( WResID *src )
 {
     WResID  *dest;
-    int     len;
+    size_t  len;
 
     if( src == NULL ) {
         return( NULL );
